+++ src/api/account.js
@@ -110,21 +110,15 @@ export async function executeOperation(
     },
   });
 
-  const arbitrumParams = chainId.startsWith('arbitrum') ? {
-    paymasterContext: { sponsorshipPolicyId },
-    userOperation: {
-      estimateFeesPerGas: async () => {
-        return (await pimlicoClient.getUserOperationGasPrice()).standard;
-      },
-    },
-  } : {};
+  const customParams = availableNetworks[chainId].getCustomPaymasterParams ? 
+    availableNetworks[chainId].getCustomPaymasterParams(pimlicoClient, sponsorshipPolicyId) : {};
 
   const smartAccountClient = createSmartAccountClient({
     account,
     chain,
     bundlerTransport: http(bundler),
     paymaster: pimlicoClient,
-    ...arbitrumParams,
+    ...customParams,
   });
 
   const txHash = await smartAccountClient.sendTransaction(callData);
