+++ src/api/params.js
@@ -17,6 +17,14 @@ export const availableNetworks = {
     juryRole: '0x9f70476b4563c57c3056cc4e8dffc8025828c99ea7a458e33c1502f84b53cc94',
     attestationNft: '0xdCa6d6E8f4E69C3Cf86B656f0bBf9b460727Bed9',
     participationNft: '0x9297845e37731480a090dB0d8eA2e2c65133523e',
+    getCustomPaymasterParams: (pimlicoClient, policyId) => ({
+      paymasterContext: { sponsorshipPolicyId: policyId },
+      userOperation: {
+        estimateFeesPerGas: async () => {
+          return (await pimlicoClient.getUserOperationGasPrice()).standard;
+        },
+      },
+    }),
   },
   'arbitrum': {
     chain: arbitrum,
@@ -28,6 +36,14 @@ export const availableNetworks = {
     juryRole: '0x9f70476b4563c57c3056cc4e8dffc8025828c99ea7a458e33c1502f84b53cc94',
     attestationNft: '0xF81508fC99Ffcfbbb5421150785c9820F8cBA9b2',
     participationNft: '0x9297845e37731480a090dB0d8eA2e2c65133523e',
+    getCustomPaymasterParams: (pimlicoClient, policyId) => ({
+      paymasterContext: { sponsorshipPolicyId: policyId },
+      userOperation: {
+        estimateFeesPerGas: async () => {
+          return (await pimlicoClient.getUserOperationGasPrice()).standard;
+        },
+      },
+    }),
   },
   'base-sepolia': {
     chain: baseSepolia,
